fixed bug for compute_pi_euler()

+++ computepi.h
@@ -8,5 +8,5 @@ double compute_pi_leibniz(size_t N);
 double compute_pi_leibniz_openmp(size_t N, int threads);
 double compute_pi_leibniz_avx(size_t N);
 double compute_pi_leibniz_avx_unroll(size_t N);
-long double compute_pi_euler(size_t N);
+double compute_pi_euler(size_t N);
